연산자와 등호에 대한 이벤트 핸들러를 상위 컴포넌트로 이동
