+++ src/App.tsx
@@ -3,7 +3,10 @@ import {useState} from 'react';
 import Digits from './components/Digits';
 import Modifiers from './components/Modifiers';
 import Operations from './components/Operations';
+
 import {Validation} from './utils/Validation';
+import {Calculator} from './utils/Calculator';
+
 import fixtures from './fixtures';
 
 function App() {
@@ -11,6 +14,8 @@ function App() {
 	const [secondNumber, setSecondNumber] = useState<string>('');
 	const [operator, setOperator] = useState<string>('');
 	const [resultNubmber, setResultNumber] = useState<string>('');
+
+	const calculator = new Calculator();
 
 	let isValidNumber = true;
 
@@ -70,6 +75,50 @@ function App() {
 		}
 	};
 
+	const handleInfinity = (result: string) => result === 'Infinity'
+		? '오류' : result;
+
+	const handleClickOperator = (value: string) => {
+		setOperator(value);
+	};
+
+	const handleClickEqualSign = () => {
+		if (secondNumber === '') {
+			setResultNumber(firstNumber);
+			setFirstNumber('');
+			setSecondNumber('');
+			setOperator('');
+			return;
+		}
+
+		let calculatedResult = 0;
+
+		const num1 = Number(firstNumber);
+		const num2 = Number(secondNumber);
+
+		switch (operator) {
+			case '+':
+				calculatedResult = calculator.sum(num1, num2);
+				break;
+			case '-':
+				calculatedResult = calculator.subtract(num1, num2);
+				break;
+			case 'X':
+				calculatedResult = calculator.multiply(num1, num2);
+				break;
+			case '/':
+				calculatedResult = calculator.divide(num1, num2);
+				break;
+			default:
+				break;
+		}
+
+		setResultNumber(handleInfinity(String(calculatedResult)));
+		setFirstNumber('');
+		setSecondNumber('');
+		setOperator('');
+	};
+
 	return (
 		<div id='app'>
 			<div className='calculator'>
@@ -79,15 +128,12 @@ function App() {
 				<Digits
 					handleClickNumber={handleClickNumber}
 				/>
-				<Modifiers onClickAc={handleClickAc} />
+				<Modifiers
+					handleClickAc={handleClickAc}
+				/>
 				<Operations
-					firstNumber={firstNumber}
-					secondNumber={secondNumber}
-					operator={operator}
-					setResultNumber={setResultNumber}
-					setFirstNumber={setFirstNumber}
-					setSecondNumber={setSecondNumber}
-					setOperator={setOperator}
+					handleClickOperator={handleClickOperator}
+					handleClickEqualSign={handleClickEqualSign}
 				/>
 			</div>
 		</div>
